requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
pillow==11.1.0
python-dotenv==1.0.1
python-linkedin==4.1
requests==2.32.3
requests-oauthlib==2.0.0
tweepy==4.15.0
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
